Adding Christina's changes to the master
Auto merge did not work. Here is Christina's description:

So the edit code took a lot longer than expected :-P I think this
should be it for the main functionality of the views. Let me know if
this doesn't merge--I think Brian has an update to the target email
pending as well.

There was a lot of repeat code for creating new terms so I went ahead
and created a createNewTerm function to clean that up a bit.