Update views.py
Added many comments in the edit view. Also added None types term1, term2, term3--if a 2nd or 3rd term was not stored in the database already and the user updated the terms list to include additional terms, an error would occur due to term2 and/or term3 not being created.