Update views.py
Edit fields are now populated with all current info except the picture.
NOTE: there IS a bug here--if you don't choose a picture, it won't recognize the form as valid. To test the edit function for the rest of the fields, choose a picture or your other data won't ever be evaluated.